src/verification.rs
#![no_std]

use soroban_sdk::{contractimpl, Address, Env};

// Importing necessary modules and types
use crate::types::{Verifier, VerificationData};
use crate::errors::VerificationError;

/// The Verification module is responsible for handling the verification process for emissions reductions.
pub struct Verification;

/// Implementation of the Verification module functionalities.
impl Verification {
    /// Verifies a project and records the verification data.
    /// This function is called by authorized verifiers to verify emissions reductions.
    pub fn verify_project(env: Env, verifier: Address, data: VerificationData) -> Result<(), VerificationError> {
        // Ensure the verifier is authorized
        Self::authorize_verifier(&env, &verifier)?;

        // Logic to verify the project
        // This could involve checking the data against certain criteria
        // and storing the verification result in the blockchain storage.

        Ok(())
    }

    /// Records the verification data on the blockchain.
    /// This function is called after a project has been successfully verified.
    pub fn record_verification(env: Env, verifier: Address, data: VerificationData) -> Result<(), VerificationError> {
        // Ensure the verifier is authorized
        Self::authorize_verifier(&env, &verifier)?;

        // Logic to record the verification data
        // This could involve storing the data in the blockchain storage
        // and emitting an event for the recorded verification.

        Ok(())
    }

    /// Authorizes a verifier to verify projects.
    /// This function checks if the verifier has the necessary permissions.
    fn authorize_verifier(env: &Env, verifier: &Address) -> Result<(), VerificationError> {
        // Logic to check if the verifier is authorized
        // This could involve checking a list of authorized verifiers stored in the blockchain.

        Ok(())
    }
}
